feat: Enhance telemetry fault logging and drive control mechanisms
- Added LogNewFaults function to notify new faults via RasLink_Log in telemetry.c, ensuring timely fault reporting.
- Introduced logged_faults in telemetry.h to track previously logged faults.
- Updated Telemetry_Update to call LogNewFaults, reducing polling frequency impact on fault detection.
- Implemented wheel lift guard in drive.c to manage torque limits based on wheel speed anomalies, enhancing traction control.
- Added functions to enable/disable traction control and wheel lift guard in drive.h.
- Modified IsAutoStopObstacleAhead to use actual vehicle speed for obstacle detection, improving safety during maneuvers.
- Enhanced IMU recovery handling to prevent false timeouts during I2C recovery.
- Updated range sensor logic to filter new measurements based on sequence numbers, improving distance accuracy.

+++ Program/MainF446RE_V3/lib/filter/lpf.h
@@ -1,20 +1,22 @@
 #ifndef LPH_H_
 #define LPH_H_
 
+// Cortex-M4F の FPU は単精度のみのため float で持つ (double だとソフトウェア
+// エミュレーションになり、毎周期何度も呼ばれるこのフィルタでは無視できないコストになる)
 typedef struct {
-      double current_val;
-      double prev_val;
-      double k_lpf;  // ローパスフィルタ係数
+      float current_val;
+      float prev_val;
+      float k_lpf;  // ローパスフィルタ係数
 } LPF;
 
-static inline void LPF_Init(LPF *lpf, double k_lpf, double initial_val) {
+static inline void LPF_Init(LPF *lpf, float k_lpf, float initial_val) {
       lpf->k_lpf = k_lpf;
       lpf->prev_val = initial_val;
 }
 
-static inline double LPF_Update(LPF *lpf, double new_val) {
+static inline float LPF_Update(LPF *lpf, float new_val) {
       // ローパスフィルタの更新
-      lpf->current_val = lpf->k_lpf * lpf->prev_val + (1.0 - lpf->k_lpf) * new_val;
+      lpf->current_val = lpf->k_lpf * lpf->prev_val + (1.0f - lpf->k_lpf) * new_val;
       lpf->prev_val = lpf->current_val;
       return lpf->current_val;
 }
